api/contacts/pyproject.toml
[tool.poetry]
name = "contacts"
version = "0.1.0"
description = ""
authors = ["Yuto Nagano <chamelion@hotmail.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.35.4"
botocore = "^1.35.4"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
alembic = "^1.13.3"
types-psycopg2 = "^2.9.21.20240819"

[tool.poetry-plugin-lambda-build]
package-artifact-path = "dist/package.zip"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
